% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vc_models.R
\name{mixed_model_pair}
\alias{mixed_model_pair}
\title{Mixed Effect Model with Pair Identifier}
\usage{
mixed_model_pair(data)
}
\arguments{
\item{data}{A list containing data from the \code{sim_tc_cv} function.}
}
\value{
A tibble with model results.
}
\description{
Internal function that fits a mixed-effect model using a pair identifier.
}
\keyword{internal}
