% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vc_models.R
\name{mixed_model_treatment_only}
\alias{mixed_model_treatment_only}
\title{Mixed Effect Model with Treatment-only Group Identifier}
\usage{
mixed_model_treatment_only(data)
}
\arguments{
\item{data}{A list containing data from the \code{sim_tc_cv} function.}
}
\value{
A tibble with model results.
}
\description{
Internal function that fits a mixed-effect model using a treatment-only group identifier.
}
\keyword{internal}
