% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vc_models.R
\name{paired_t_test}
\alias{paired_t_test}
\title{Paired T-test Comparison}
\usage{
paired_t_test(data)
}
\arguments{
\item{data}{A list containing data from the \code{sim_tc_cv} function.}
}
\value{
A tibble with t-test results.
}
\description{
Internal function that conducts a paired t-test comparison.
}
\keyword{internal}
