% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vc_models.R
\name{two_step_t_test}
\alias{two_step_t_test}
\title{Two-Step Adjusted Paired t-test}
\usage{
two_step_t_test(data)
}
\arguments{
\item{data}{A list containing data from the \code{sim_tc_cv} function.}
}
\value{
A tibble with t-test results.
}
\description{
Internal function that conducts a two-step adjusted paired t-test.
}
\keyword{internal}
