% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vc_summary.R
\name{vc_summary}
\alias{vc_summary}
\title{Summarize Results from Various Modeling Approaches}
\usage{
vc_summary(data, remove_re = TRUE, remove_int = TRUE)
}
\arguments{
\item{data}{A list containing data from the \code{sim_tc_cv} function.}

\item{remove_re}{Logical. If TRUE, removes the random effects (e.g., "sd__(Intercept)") from the summary. Default is TRUE.}

\item{remove_int}{Logical. If TRUE, removes the intercept term from the summary. Default is TRUE.}
}
\value{
A tibble summarizing results from the different modeling approaches.
}
\description{
This function takes the simulated data and summarizes results from various modeling
approaches, including mixed models and t-tests.
}
\examples{
set.seed(123)
simulated_data <- sim_tc_cv(n_treatment = 100, n_control = 1000, n_matches = 10)
summary_results <- vc_summary(simulated_data)
head(summary_results)
}
